requirements.txt
requests
numpy
pandas
scikit-learn
Pillow
matplotlib
scipy
